flatten pending and published api response
